refactor(Divider): convert to ts
