+++ docs/src/examples/components/Divider/Types/DividerExample.tsx
