+++ docs/src/examples/components/Divider/Variations/index.tsx
@@ -3,11 +3,11 @@ import ComponentExample from 'docs/src/components/ComponentDoc/ComponentExample'
 import ExampleSection from 'docs/src/components/ComponentDoc/ExampleSection'
 
 const Variations = () => (
-  <ExampleSection title='Variations'>
+  <ExampleSection title="Variations">
     <ComponentExample
-      title='Size'
-      description='A divider can have different sizes.'
-      examplePath='components/Divider/Variations/DividerExampleSize'
+      title="Size"
+      description="A divider can have different sizes."
+      examplePath="components/Divider/Variations/DividerExampleSize"
     />
   </ExampleSection>
 )
