+++ docs/src/examples/components/Divider/index.tsx
